requirements.txt
fastapi==0.110.2
numpy~=1.24.4
Pillow==10.3.0
PyYAML==6.0.1
Requests==2.31.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
uvicorn==0.29.0
scikit-image~=0.21.0
opencv-python~=4.9.0.80
pandas~=2.0.3
scikit-learn~=1.3.2
python-multipart==0.0.5